server/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
black==23.10.1
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cohere==4.32
colorama==0.4.6
dataclasses-json==0.5.14
datasets==2.14.6
demjson3==3.0.6
Deprecated==1.2.14
dill==0.3.7
diskcache==5.6.3
exceptiongroup==1.1.3
faiss-cpu==1.7.4
fastavro==1.8.2
filelock==3.13.1
FLAML==2.1.1
Flask==3.0.0
frozenlist==1.4.0
fsspec==2023.10.0
google-search-results==2.4.2
greenlet==3.0.1
h11==0.14.0
huggingface-hub==0.18.0
ibm-cos-sdk==2.13.2
ibm-cos-sdk-core==2.13.2
ibm-cos-sdk-s3transfer==2.13.2
ibm-watson-machine-learning==1.0.327
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.330
langsmith==0.0.57
llama-index==0.8.59
lomond==0.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.26.1
openai==0.28.1
outcome==1.3.0.post0
packaging==23.2
pandas==1.5.3
pathspec==0.11.2
platformdirs==3.11.0
pluggy==1.3.0
prettytable==3.9.0
prompt-toolkit==3.0.36
pyarrow==14.0.0
pyautogen==0.1.14
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pymemgpt==0.1.17
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
PyPDF2==3.0.1
PySocks==1.7.1
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
questionary==2.0.1
regex==2023.10.3
requests==2.31.0
rich==13.6.0
selenium==4.15.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.23
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
tiktoken==0.5.1
tomli==2.0.1
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
wcwidth==0.2.9
Werkzeug==3.0.1
wrapt==1.15.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
